`bundle check` does more than just check the bundle -- it can update Gemfile.lock
After a modification to 'Gemfile', guard-bundler would sometimes report 'Bundle already up-to-date' when it in fact the bundle was not up to date. It turns out that the 'bundle check' command updates 'Gemfile.lock' and returns with success when all dependencies can be met by using locally installed gems.

This PR refactors the #bundle_refresh method to handle this additional behavior of the 'bundle check' command. Tests are updated to accomodate these changes.